房间完成

+++ model/src/main/java/com/scccy/stgy/model/mapper/RoomAttrValueMapper.java
@@ -2,11 +2,10 @@ package com.scccy.stgy.model.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.scccy.stgy.model.domain.RoomAttrValue;
-
-import java.util.List;
-
 import com.scccy.stgy.model.vo.RoomAttrValueVo;
 import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
 
 public interface RoomAttrValueMapper extends BaseMapper<RoomAttrValue> {
     int updateBatch(List<RoomAttrValue> list);
@@ -17,5 +16,5 @@ public interface RoomAttrValueMapper extends BaseMapper<RoomAttrValue> {
 
     int insertOrUpdateSelective(RoomAttrValue record);
 
-    List<RoomAttrValueVo> getAttrValueVos(Integer id);
+    List<RoomAttrValueVo> getAttrValueVos(Long id);
 }
