+++ web/web-admin/src/main/java/com/scccy/stgy/web/admin/service/RoomAttrValueService.java
@@ -1,9 +1,10 @@
 package com.scccy.stgy.web.admin.service;
 
-import java.util.List;
-import com.scccy.stgy.model.domain.RoomAttrValue;
 import com.baomidou.mybatisplus.extension.service.IService;
+import com.scccy.stgy.model.domain.RoomAttrValue;
 import com.scccy.stgy.model.vo.RoomAttrValueVo;
+
+import java.util.List;
 
 public interface RoomAttrValueService extends IService<RoomAttrValue>{
 
@@ -16,5 +17,5 @@ public interface RoomAttrValueService extends IService<RoomAttrValue>{
 
     int insertOrUpdateSelective(RoomAttrValue record);
 
-    List<RoomAttrValueVo> getAttrValueVos(Integer id);
+    List<RoomAttrValueVo> getAttrValueVos(Long id);
 }
